# docs/swagger.yaml
basePath: /api
definitions:
  configs.Task:
    properties:
      completed:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      user:
        type: integer
    type: object
  configs.User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://example.com/contact
  description: API для управления пользователями и задачами
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: ToDo API
  version: "1.0"
paths:
  /login:
    post:
      description: Авторизация пользователя по email и паролю
      parameters:
      - description: Email и пароль
        in: body
        name: credentials
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      responses:
        "200":
          description: JWT токен
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неверный email или пароль
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Авторизация
      tags:
      - Авторизация
  /tasks:
    get:
      description: Возвращает список всех задач из базы данных
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/configs.Task'
            type: array
        "404":
          description: Список задач пуст
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка базы данных
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список задач
      tags:
      - Задачи
    post:
      description: Создает новую задачу для авторизованного пользователя
      parameters:
      - description: Данные задачи
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/configs.Task'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/configs.Task'
        "400":
          description: Неверный формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Не удалось создать задачу
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать задачу
      tags:
      - Задачи
  /tasks/{id}:
    delete:
      description: Удаляет задачу по её ID, если пользователь является её владельцем
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Задача удалена
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Вы не можете удалить эту задачу
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Не удалось удалить задачу
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить задачу
      tags:
      - Задачи
    get:
      description: Возвращает данные задачи по её ID
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/configs.Task'
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Задача не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка базы данных
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить задачу
      tags:
      - Задачи
    put:
      description: Обновляет данные задачи по её ID
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      - description: Данные задачи
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/configs.Task'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/configs.Task'
        "400":
          description: Неверный ID или формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Вы не можете обновить эту задачу
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Не удалось обновить задачу
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить задачу
      tags:
      - Задачи
  /users:
    post:
      description: Создает нового пользователя и сохраняет его в базе данных
      parameters:
      - description: Данные пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/configs.User'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/configs.User'
        "400":
          description: Неверный формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Не удалось создать пользователя
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать пользователя
      tags:
      - Пользователи
  /users/{id}:
    delete:
      description: Удаляет пользователя из базы данных по его ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Пользователь удален
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Не удалось удалить пользователя
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить пользователя
      tags:
      - Пользователи
    get:
      description: Возвращает информацию о пользователе по его ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/configs.User'
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка базы данных
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить пользователя
      tags:
      - Пользователи
    put:
      description: Обновляет данные пользователя по его ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Данные пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/configs.User'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/configs.User'
        "400":
          description: Неверный ID или формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Не удалось обновить пользователя
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить пользователя
      tags:
      - Пользователи
swagger: "2.0"